Refactor GMailinator.m and other fixes.
The goal here is to make easier to understand and add new shortcuts to
GMailinator. Any new addition should be straightforward on the main file.

Also I gather several fixes for different thing over everyone's fork to get the
following:

Auto UUID on the build script;
Most of the compilation warnings are gone, with exception of the deprecation to
load the NIB. If I fix that, the Move To dialog simply does not show up;

Also I got rid of a memory leak on the event creation. It as a tiny one and
would take forever to impact real users but nevertheless it was a good catch.